refactor: extract AxesTab from ChartPropertiesPanel
